+++ src/components/Select.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from "react";
 
 import classnames from "classnames";
-import { existsBy } from "neetocist";
+import { _existsBy } from "neetocist";
 import { Down, Close } from "neetoicons";
 import PropTypes from "prop-types";
 import { prop, assoc, flatten, pluck } from "ramda";
@@ -21,12 +21,13 @@ const SIZES = { small: "small", medium: "medium", large: "large" };
 const STRATEGIES = { default: "default", fixed: "fixed" };
 
 const Control = ({ children, ...props }) => {
-  const { selectProps } = props;
+  const { selectProps, hasValue } = props;
 
   return (
     <components.Control {...props}>
-      {selectProps.isMulti && (
-        <span className="neeto-ui-btn neeto-ui-btn--style-primary neeto-ui-react-select__add-btn">
+      {/* hasValue is part of commonProps passed to custom components internally by react-select */}
+      {hasValue && selectProps.isMulti && (
+        <span className="neeto-ui-btn neeto-ui-btn--style-secondary neeto-ui-react-select__add-btn">
           {selectProps.addButtonLabel || "Add"}
         </span>
       )}{" "}
@@ -45,7 +46,10 @@ const DropdownIndicator = props => (
 );
 
 const ClearIndicator = props => (
-  <components.ClearIndicator {...props}>
+  <components.ClearIndicator
+    {...props}
+    innerProps={{ ...props.innerProps, "data-cy": "clear-select-indicator" }}
+  >
     <Close size={16} />
   </components.ClearIndicator>
 );
@@ -242,7 +246,7 @@ const Select = ({
     let currentOptions = options || defaultOptions;
     if (Array.isArray(value)) value = value[0];
 
-    const isGrouped = existsBy({ options: Array.isArray }, currentOptions);
+    const isGrouped = _existsBy({ options: Array.isArray }, currentOptions);
 
     if (isGrouped) {
       currentOptions = flatten(pluck("options", currentOptions));